MakefileS + reorg + norme

+++ hash_tables/ht_free.c
@@ -6,7 +6,7 @@
 /*   By: ggilaber <ggilaber@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/01/27 15:53:48 by ggilaber          #+#    #+#             */
-/*   Updated: 2016/02/06 16:37:52 by ggilaber         ###   ########.fr       */
+/*   Updated: 2016/02/12 19:12:43 by ggilaber         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 
 void	ht_free(t_hash_tbl *ht)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < ht->size)
