+++ hash_tables/ht_freem.c
@@ -6,7 +6,7 @@
 /*   By: ggilaber <ggilaber@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/01/27 15:53:48 by ggilaber          #+#    #+#             */
-/*   Updated: 2016/02/06 19:14:44 by ggilaber         ###   ########.fr       */
+/*   Updated: 2016/02/12 19:15:01 by ggilaber         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ht_freem(t_hash_tbl *ht)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < ht->size)
